Change win check algorithm. Now checks diagonals as well.
Algorithm iterates over every square on the board. If the square is not 0, it attempts to iterate three additional steps in each of the eight possible directions. If it completes the three additional steps without index going out of bounds as well as not finding a value different than the original square, a four in a row has been found.